requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2.13
django-environ==0.8.1
django-localflavor==3.1
django-qrcode==0.3
django-simple-history==3.1.1
django-widget-tweaks==1.4.12
gunicorn==20.1.0
idna==3.3
install==1.3.5
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.4.0
python-stdnum==1.17
pytz==2021.3
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
twilio==7.9.3
urllib3==1.26.9
whitenoise==5.3.0
